♻️ update: (tabs) 사이드바 라우트 그룹 재구성
