+++ packages/frontend/app/(tabs)/(sidebar)/dashboard/components/interview-start-box.tsx
