+++ packages/frontend/app/(tabs)/(sidebar)/dashboard/page.tsx
@@ -7,7 +7,7 @@ import { getUserSession } from "@/app/lib/server/session";
 import InterviewList from "./components/interview-list";
 import InterviewStartBox from "./components/interview-start-box";
 import InterviewWelomeHeader from "./components/interview-welcome-header";
-import Tip from "../(simulator)/interview/[id]/result/components/tip";
+import Tip from "@/app/(tabs)/(sidebar)/interview/[id]/result/components/tip";
 
 export const dynamic = "force-dynamic";
 
@@ -25,25 +25,23 @@ export default async function Page() {
   const interviews = await getInterviews();
 
   return (
-    <div>
-      <main className="mx-auto max-w-180 px-5">
+    <section className="mx-auto w-full max-w-3xl px-5 pb-5">
+      <div className="mt-12">
+        <InterviewWelomeHeader username={user.email} />
+      </div>
+      <div className="mt-8">
+        <InterviewStartBox href="/interview/create" />
+      </div>
+      {interviews.length > 0 && (
         <div className="mt-12">
-          <InterviewWelomeHeader username={user.email} />
+          <InterviewList interviews={interviews} />
         </div>
-        <div className="mt-8">
-          <InterviewStartBox href="/interview/create" />
+      )}
+      <div className="mt-12">
+        <div className="rounded-2xl bg-primary/10 p-5">
+          <Tip />
         </div>
-        {interviews.length > 0 && (
-          <div className="mt-12">
-            <InterviewList interviews={interviews} />
-          </div>
-        )}
-        <div className="mt-12">
-          <div className="rounded-2xl bg-primary/10 p-5">
-            <Tip />
-          </div>
-        </div>
-      </main>
-    </div>
+      </div>
+    </section>
   );
 }
