+++ packages/frontend/app/(tabs)/(sidebar)/documents/components/ocr-info-box.tsx
