+++ packages/frontend/app/(tabs)/(sidebar)/interview/[id]/ready/layout.tsx
