+++ packages/frontend/app/(tabs)/(sidebar)/interview/[id]/result/components/ai-summary.tsx
