+++ packages/frontend/app/(tabs)/(sidebar)/interview/[id]/result/components/panel.tsx
