+++ packages/frontend/app/(tabs)/(sidebar)/interview/[id]/result/components/recent-recording.tsx
