+++ packages/frontend/app/(tabs)/(sidebar)/interview/[id]/result/components/score.tsx
