+++ packages/frontend/app/(tabs)/(sidebar)/interview/[id]/result/components/tip.tsx
