+++ packages/frontend/app/(tabs)/(sidebar)/interview/[id]/result/schema.ts
