+++ packages/frontend/app/(tabs)/(simulator)/interview/[id]/(simulator)/layout.tsx
@@ -1,0 +1,3 @@
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return children;
+}
