+++ packages/frontend/app/(tabs)/(simulator)/layout.tsx
@@ -1,5 +1,3 @@
-import Header from "@/app/components/header";
-
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <div className="flex h-dvh flex-col">
